Arrange Array - Leetcode Medium Q with effective soln FirstUnique - Leetcode Easy Q ArrayMedian - Q

+++ src/com/leetcode/ArrayMedian.java
@@ -34,7 +34,7 @@ public class ArrayMedian {
                 if(x+y%2 == 0){
                     return ((double)Math.max(maxLeftX, maxLeftY)+Math.min(minRightX, minRightY))/2;
                 }else {
-                    return (double) Math.max(maxLeftX, maxLeftY);
+                    return (Math.max(maxLeftX, maxLeftY));
                 }
             }else if(maxLeftX>=minRightY){
                 hi = pX-1;
